Refactor some common autotune-related utils into a new file
Summary: I'll need some of the benchmark-related functions surfaced in a common utils file so I can use them for remote autotuning. This PR is almost a straight move of a few nested functions from inside AlgorithmSelectorCache, plus the datatypes they depend on, except:
* All hell breaks loose with mypy on select_algorithm.py with this refactor, so I disabled it. Initial investigations looks like there are potentially a ton of type problems in this file, so I'll tackle fixing types in a follow-up diff.
* Introduced a new interface "Benchmarkable" as the base type that gets benchmarked. This is looking forward to the remote autotuning implementation where the types that we'll send across the wire must be serializable. So we won't be using the exact same types we pass to these utilities now, e.g., ExternKernelCaller.
* Rather than checking types explicitly, I introduced an is_extern() method to the Benchmarkable interface.

Test Plan: Existing unit tests

[ghstack-poisoned]
